+++ public_data/serializers.py
@@ -161,3 +161,16 @@ class ZoneUrbaSerializer(serializers.GeoFeatureModelSerializer):
         )
         model = models.ZoneUrba
         geo_field = "mpoly"
+
+
+class SearchLandSerializer(serializers.Serializer):
+    needle = serializers.CharField(required=True)
+
+
+class LandSerializer(serializers.Serializer):
+    id = serializers.IntegerField()
+    name = serializers.CharField()
+    source_id = serializers.SerializerMethodField()
+
+    def get_source_id(self, obj) -> str:
+        return obj.get_official_id()
