+++ public_data/urls.py
@@ -1,49 +1,31 @@
 from django.urls import path
 from rest_framework import routers
 
-from public_data.api.search import SearchLandApiView
-
-from .api.views import (
-    ArtificialAreaViewSet,
-    CommuneViewSet,
-    CouvertureSolViewset,
-    DepartementViewSet,
-    EpciViewSet,
-    OcsgeDiffCentroidViewSet,
-    OcsgeDiffViewSet,
-    OcsgeViewSet,
-    RegionViewSet,
-    ScotViewSet,
-    UsageSolViewset,
-    ZoneConstruiteViewSet,
-    ZoneUrbaViewSet,
-    grid_views,
-)
-from .views import DisplayMatrix
+from public_data import views
 
 app_name = "public_data"
 
 
 urlpatterns = [
-    path("matrix", DisplayMatrix.as_view(), name="matrix"),
-    path("grid", grid_views, name="grid"),
-    path("search-land", SearchLandApiView.as_view(), name="search-land"),
+    path("matrix", views.DisplayMatrix.as_view(), name="matrix"),
+    path("grid", views.grid_views, name="grid"),
+    path("search-land", views.SearchLandApiView.as_view(), name="search-land"),
 ]
 
 
 router = routers.DefaultRouter()
-router.register(r"communes", CommuneViewSet)
-router.register(r"referentiel/couverture-sol", CouvertureSolViewset)
-router.register(r"referentiel/usage-sol", UsageSolViewset)
-router.register(r"ocsge/general", OcsgeViewSet)
-router.register(r"ocsge/diff", OcsgeDiffViewSet)
-router.register(r"ocsge/diff-centroids", OcsgeDiffCentroidViewSet, basename="ocsgeDiffCentroids")
-router.register(r"ocsge/zones-construites", ZoneConstruiteViewSet)
-router.register(r"ocsge/zones-artificielles", ArtificialAreaViewSet)
-router.register(r"referentiel/region", RegionViewSet)
-router.register(r"referentiel/departement", DepartementViewSet)
-router.register(r"referentiel/epci", EpciViewSet)
-router.register(r"referentiel/scot", ScotViewSet)
-router.register(r"referentiel/zones-urbaines", ZoneUrbaViewSet)
+router.register(r"communes", views.CommuneViewSet)
+router.register(r"referentiel/couverture-sol", views.CouvertureSolViewset)
+router.register(r"referentiel/usage-sol", views.UsageSolViewset)
+router.register(r"ocsge/general", views.OcsgeViewSet)
+router.register(r"ocsge/diff", views.OcsgeDiffViewSet)
+router.register(r"ocsge/diff-centroids", views.OcsgeDiffCentroidViewSet, basename="ocsgeDiffCentroids")
+router.register(r"ocsge/zones-construites", views.ZoneConstruiteViewSet)
+router.register(r"ocsge/zones-artificielles", views.ArtificialAreaViewSet)
+router.register(r"referentiel/region", views.RegionViewSet)
+router.register(r"referentiel/departement", views.DepartementViewSet)
+router.register(r"referentiel/epci", views.EpciViewSet)
+router.register(r"referentiel/scot", views.ScotViewSet)
+router.register(r"referentiel/zones-urbaines", views.ZoneUrbaViewSet)
 
 urlpatterns += router.urls
